% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qdata.R
\name{get_qdata}
\alias{get_qdata}
\title{Getting quarterly data from all public companies in EDGAR}
\usage{
get_qdata(account = "Revenues", years = 2020:2023, quarters = c("Q3"))
}
\arguments{
\item{account}{A string representing the account.}

\item{years}{A sequence of numeric values representing the years.}

\item{quarters}{A string representing the quarter.}
}
\value{
A dataframe
}
\description{
Getting quarterly data from all public companies in EDGAR
}
