% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Data wrangling for tidy fundamental data from EDGAR}
\usage{
prepare_data(df = NULL, quarterly = TRUE, ...)
}
\arguments{
\item{df}{A dataframe.}

\item{quarterly}{Boolean indicating if quarterly data is present.}

\item{...}{Additional dataframes passed.}
}
\value{
A dataframe
}
\description{
Data wrangling for tidy fundamental data from EDGAR
}
\examples{
prepare_data(get_qdata(account = "NetIncomeLoss", years = 2022:2023, quarters = c("Q4")))
}
