% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_qdata}
\alias{prepare_qdata}
\title{Data wrangling for tidy fundamental data from EDGAR}
\usage{
prepare_qdata(df = NULL, ...)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Additional dataframes passed.}
}
\value{
A dataframe
}
\description{
Data wrangling for tidy fundamental data from EDGAR
}
