% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Helper function for quarterly financial data retrieval}
\usage{
retrieve_data(account, year, quarter, taxonomy = "us-gaap", unit = "USD")
}
\arguments{
\item{account}{A string representing the account.}

\item{year}{A numeric value representing the year.}

\item{quarter}{A string representing the quarter.}

\item{taxonomy}{A string representing the taxonomy.}

\item{unit}{A string representing the units.}
}
\value{
A dataframe
}
\description{
Helper function for quarterly financial data retrieval
}
