% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_max.R
\name{safe_max}
\alias{safe_max}
\title{Safely calculating the max.}
\usage{
safe_max(x, na.rm = FALSE)
}
\arguments{
\item{x}{A number.}

\item{na.rm}{Boolean.}
}
\value{
A number.
}
\description{
Safely calculating the max.
}
