% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_data.R
\name{yearly_data}
\alias{yearly_data}
\title{Getting a summary with the basic financials for all companies}
\usage{
yearly_data(years = 2020:2023)
}
\arguments{
\item{years}{A sequence of numeric values representing the years.}
}
\value{
A dataframe
}
\description{
Getting a summary with the basic financials for all companies
}
\examples{
\donttest{yearly_data(years = 2022:2023)}
}
